Se modificaron las clases del modelo añadiendoles un nuevo atributo, además, se implemento autenticación y permisos de acceso por rol

+++ src/main/java/com/cristian/helpdesk/helpdesk/service/UserService.java
@@ -1,10 +1,16 @@
 package com.cristian.helpdesk.helpdesk.service;
 
+import com.cristian.helpdesk.helpdesk.model.Role;
 import com.cristian.helpdesk.helpdesk.model.Ticket;
 import com.cristian.helpdesk.helpdesk.model.User;
 import com.cristian.helpdesk.helpdesk.repository.TicketRepository;
 import com.cristian.helpdesk.helpdesk.repository.UserRepository;
+import com.cristian.helpdesk.helpdesk.security.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.Collections;
@@ -20,8 +26,18 @@ public class UserService {
     @Autowired
     private TicketRepository ticketRepository;
 
+    @Autowired
+    private SecurityUtils securityUtils;
+
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
     public User saveUser(User user) {
-        return userRepository.save(user);
+        if (!userRepository.existsById(user.getNit()) && securityUtils.isAdmin()) {
+            user.setPassword(passwordEncoder.encode(user.getPassword()));
+            return userRepository.save(user);
+        }
+        throw new IllegalArgumentException("La cédula se encuentra registrada");
     }
 
     public List<User> listUser() {
@@ -33,7 +49,9 @@ public class UserService {
     }
 
     public void deleteUser(String cedula) {
-        userRepository.deleteById(cedula);
+        if (securityUtils.isAdmin()) {
+            userRepository.deleteById(cedula);
+        }
     }
 
     public boolean existEmail(String email) {
@@ -42,17 +60,29 @@ public class UserService {
 
 
     public Optional<User> updateUser(User user) {
-        Optional<User> existingUser = userRepository.findById(user.getCedula());
+        Optional<User> existingUser = userRepository.findById(user.getNit());
         if (existingUser.isPresent()) {
-            return Optional.of(userRepository.save(user));
-        } else {
-            return Optional.empty();
+            User aux = existingUser.get();
+            if (securityUtils.getEmailAuth().equals(aux.getEmail()) || securityUtils.isAdmin()) {
+                aux.setName(user.getName());
+                aux.setPassword(passwordEncoder.encode(user.getPassword()));
+                aux.setRole(user.getRole());
+                aux.setEmail(user.getEmail());
+                aux.setNit(user.getNit());
+                return Optional.of(userRepository.save(user));
+            }
         }
+        return Optional.empty();
     }
 
     public List<Ticket> getTicketsByUserCedula(String cedula) {
         Optional<User> user = userRepository.findById(cedula);
-        return user.map(User::getTickets).orElse(Collections.emptyList());
+        if (user.isPresent()) {
+            if (!securityUtils.isClient() || securityUtils.getEmailAuth().equals(user.get().getEmail())) {
+                return user.map(User::getTickets).orElse(Collections.emptyList());
+            }
+        }
+        return Collections.emptyList();
     }
 
     public Optional<Ticket> addTicketUser(String cedula, int id) {
